text: Switch to weston_client_start instead of weston_client_launch
We really don't want to deal with sighups and pids.  It's far easier
to just deal with the client destroyed signal to respawn the input method.

Inspiration taken from Pekka's commit 826dc14ec43b92d

This fixes a potential crash on shutdown.  If the input method client
pointer is still set (sighup handler hasn't run) but the client is
already destroyed, a call to text_backend_destroy() will try to destroy
the client a second time resulting in a segfault.

Signed-off-by: Derek Foreman <derekf@osg.samsung.com>
Reviewed-by: Bryce Harrington <bryce@osg.samsung.com>